when a challenge fails, purge the challenge from the local cache of challenges so that on the next run the user can get a new one and try again
A challenge fails probably because the local machine and ACME see something different on the network, e.g. DNS propagation is in progress. (Otherwise the local test would fail and we wouldn't even try to submit the challenge.) But we were caching the invalid challenge, and so each subsequent run would report the same failure. Now we report once and drop the challenge from the cache.